+++ legate/CMakeLists.txt
@@ -36,4 +36,4 @@ legate_add_cpp_subdirectory(cpp TARGET legate_kvikio EXPORT legate_kvikio-export
 legate_default_python_install(legate_kvikio EXPORT legate_kvikio-export)
 
 # Generates `install_info.py`
-legate_add_cffi(${CMAKE_SOURCE_DIR}/cpp/legate_kvikio.hpp TARGET legate_kvikio)
+legate_add_cffi(${CMAKE_SOURCE_DIR}/cpp/task_opcodes.hpp TARGET legate_kvikio)
